+++ src/main/java/org/deepsymmetry/beatlink/DeviceAnnouncementListener.java
@@ -18,7 +18,7 @@ public interface DeviceAnnouncementListener {
      *
      * <p>Device announcements are delivered to listeners on the
      * <a href="https://docs.oracle.com/javase/tutorial/uiswing/concurrency/dispatch.html">Event Dispatch thread</a>,
-     * so it is fine to interact with user interface objects in listener methods. Any code in the listener method
+     * so it is fine to interact with user interface objects in this method. Any code in this method
      * must finish quickly, or unhandled events will back up and the user interface will be come unresponsive.</p>
      *
      * @param announcement the message which announced the device's presence
@@ -30,7 +30,7 @@ public interface DeviceAnnouncementListener {
      *
      * <p>Device announcements are delivered to listeners on the
      * <a href="https://docs.oracle.com/javase/tutorial/uiswing/concurrency/dispatch.html">Event Dispatch thread</a>,
-     * so it is fine to interact with user interface objects in listener methods. Any code in the listener method
+     * so it is fine to interact with user interface objects in this method. Any code in this method
      * must finish quickly, or unhandled events will back up and the user interface will be come unresponsive.</p>
      *
      * @param announcement the last message which was sent by the device before it disappeared